Fixed missing IdManager, improved delete performance.
Delete performance requires that each node is found, since we need to know the child nodes, we could improve this by creating a list of node id's in the hierarchy but that might overflow the SQL statement limits. If delete performance becomes a real problem there is some work to do in this area.



git-svn-id: https://source.sakaiproject.org/contrib/tfd/trunk/hierarchy@2054 fdecad78-55fc-0310-b1b2-d7d25cf747c9