WL-1006 Removed Cambridge specific references from message properties.
This consisted of the following:
- 1 reference to 'Sakai'.
- 1 reference to 'Raven' login.
- 12 references to 'CamTools'.
In the latter scenario, 11 of these occurances existed as redundant prefixes to
'site'. In these cases what I did was to bump such name-value pairs into the
default (en_US) resource bundle and removed the 'CamTools'. I could have just
deleted them from the en_GB bundle, but in each case the wording seemed to be
better / clearer than what was in the en_US version!

Squased commit:
Removed further Cambridge specific references in messages: 'Tripos'.

According to the OED, Tripos has *2* meanings specific to Cambridge University!
1 has an Oxford University equivalent - Terræ Filius. The other (which I think
was probably the way there were using it) is 'The list of candidates qualified
for the honour degree in mathematics'. This has no Oxford equivalent ;-).