WL-1074 Removed the repeated instructions from the survey dashboard.
I didn't fully investigate what changed when, but when I compared the properties
that rendered the corresponding messages, it appears that the en_US and en_GB
values had got out of step with one another. By harmonising the en_GB values
with the en_US ones the repeated message dissolved away.