Switch to joining departments -> divisions on id.
This bug was caused by the database refector that happened, new columns were introduced and the old ones left around but not updated. We weren't told about the change so continued to use the old columns and as a result didn't find any of the new/updated data.

Also switched to using logging in this commit, e.printStackTrace() shouldn't ever make it into production.