Disable links to the current page in the tool.
Also remember the last page we were on so when you click the cancel button on pages such as the new exam paper we go back to the last page. We also remember page state so if you had filtered the list of papers we return to the list of exam papers with the filter still applied.

Only essential pages should remember the previous page as otherwise you end up with a chain of pages that's can't be GCed. Could always looks at SoftReferences.