WL-1862 Allow upload to work when no papers/
If an uploaded zip file doesn't have a papers/ folder in it then the upload fails. Windows doesn't seem to create folders and relies in the extracting program creating the parent folders.

The fix involved just checking that we have entries which have the desired prefix. So an entry of papers/1234.pdf is enough.