WL-1848 Correct schema for Hibernate Startup.
Basically the groupId column in the course_group_category had become 256 characters and Hibernate creates all columns longer than 255 characters as TEXT and you can't have a index on a TEXT column. The reason for this restriction is even if you have a VARCHAR of more than 255 characters MySQL only index 255 characters.

The reason for the index getting created was the reference to the table by course_group which was creating the foreign key and hence the index.