Refactored hierarchy so we have redirect nodes.
Here we have added support for different node types in the hierarchy (Sites and Redirect).
Sites are nodes that contain a site and when visiting that node you are shown the site.
Redirect nodes just redirect you to another URL.

This is to allow sites to be moved in the hierarchy and then redirect put in to keep old links working.
There is no support for creating redirect nodes yet.