You can't return a full ModelMap when validating.
If you have a method returning a ModelMap annotated with a @ModelAttribute then it working during standard requests, but when bean validation is done as part of the request the ModelMap returned from the method is thrown away (from testing).

The workaround seems to be to just ask for the existing model to be passed into your @ModelAttribute annotated method instead and add stuff to it.