Don't call directly to Sakai from JAX-RS resource.
We use the SakaiProxy to contain all references to Sakai code and here we needed to know if the request was authenticated so we added a new method to the SakaiProxy and switched the SignupResource to use that new method.

We also extracted a method to reduce the amount of code contained here.