Using maps doesn't work with an static serializer.
Our exceptions just had an entity of a map which would fail the serialize when the object mapper was configured to use static type. The current reason for using static types is that we want to ignore any hibernate values on the beans.

In future it may well be better to explicitly support hibernate through. https://github.com/FasterXML/jackson-datatype-hibernate