Correctly setup the signup components.
This looks to be the wrong way around. I would have expected it to be the signup that handles the components that it is signed up to. The possible reason for this is that the components need updating because if they are accepted the places changes, but this normally happens in a different transaction.