WL-2978 Get factory from simple method.
The newFactory(String, Classloader) works differently across JDK6/JDK7, switching to the no-arg newFactory(). Under JDK6 it takes the name of the class to create a new instance of, under JDK7 it takes the name of the factory. I haven't tested under JDK5 as we no longer need to support that.