WL-3479 Rework date picker to use new one.
This creates an RSF evolver so that the date component is re-useable and the model doesn’t have to be altered to support the new format that comes back from the request.

The evolver is going back into mainline RSF but as we don’t want to wait for that to happen and we also aren’t yet on the latest version of RSF we are putting the code directly into the evaluations tool.