WL-3575 Correctly find components.xml
When this project changed from being an indie (build on it’s own) to being part of the standard Sakai build it started inheriting master/pom.xml which redefines where the component.xml should be found. The sakai plugin doesn’t fail when components.xml can’t be found so it just built a component without a startup file.

We just re-define it back to be a sensible maven default.