WL-1742 Don’t display options that make no sense.
When a user doesn’t have to be logged in to take a survey they can’t sensibly come back to the survey and change it. Also when you aren’t required to login you don’t know the role a person is in so there’s no point in having an option for it.

I didn’t collapse the 2 statements down into one to make it a little more readable.