WL-3721 Don’t fall over when session is invalid.
The problem was that a session didn’t have any dates and this was causing the populator to fall over with a NPE. Now we validate the sessions and drop them if they are not valid. As sessions don’t seems to have good IDs (unique in document), we logs the errors against the presentation containing them.