WL-3773 Upgrade to raven-sentry (5.0.2)
This is needed because the local sentry service is going to be upgraded to sentry 7. This means it will only support versions 5 and 6 of the protocol. We can’t upgrade to raven-sentry 6 as if we go live before sentry.oucs.ox.ac.uk has been upgraded we would no longer be able to log messages.