WL-3830 Make flushes happen sooner to stop losing data.
We have been seeing problems where the collections of Strings have been lost when the importer is running. It seems to be related to how the production code uses transactions as I was unable to reproduce this in a test.

Due to having spend too much time on this already the simple fix is to just get the code to write through the DB early which seems to fix it issue.

This might also fix the problems with having multiple importers running concurrently against a clean DB, but I haven’t checked this.