WL-3010 Don’t display an ugly message when dealt with.
If a signup has already been dealt with we tell the administrator this, rather than showing an ugly error message. To know if something has already been dealt with we have a graph of status progressions. This is so we can easily know the previous statuses a signup might have been in.