Remove duplicate servlet that was not needed.
We had 2 copies of the Jersey Application starting up, this was so that one could have basic auth in front of it. However the basic auth is just done in a filter so we can mount the same servlet at 2 URLs and just put the basic auth filter in front of them.

Also removed commented out sections and added a few comments.