WL-4038 Filter on teaching starts date.
The filtering of the signups wasn’t working, but when we cleaned up the slots/teaching-details we accidentally had the filter populated. But it filtered the signups down to only those with teaching details and there wasn’t any way to show all the signups.

This adds a way to show all the signups (default) but also filter based on when the item starts.