WL-4029 Fix Basic Auth for export.
Fix the handling of basic auth so that it’s checked on all requests and if it’s present it’s used and if it’s not present we just carry on. This allows simplification of servlets paths.

Basic Auth was broken because the servlet path based filter was getting applied before the servlet name based filters are and the authentication was lost because the request hadn’t been correctly setup.