Get the integration tests working again.
This allows them to easily be used from both Maven and the IDE.
The only bad thing is that if the test are being run then the test.properties file in bundled into the artifact.

Having the properties managed by maven makes it easier to keep the configuration outside the artifacts but it means the IDE can’t always easily run the tests.