Rework of email sending.
This extracts out the sending of emails so that it’s done on a rule based approach. If we’re going todo calendar announcements then we need to have more testable separate emails. There could do with being a re-work of the email sending methods, but I need to get back to actually adding features rather than just improving the codebase.