WL-3975 Return the minimum of data for admin list.
When getting a list of course groups that a user is an administrator for we don’t need all the nested information, just the high level object. Returning everything causes performance problems for users who are administrators for lots of groups.

This is a generic way of supporting filtering the JSON output that we generate and starts the transition to the use JAX-RS framework as it was designed, rather than directly talking to the ObjectMapper.