WL-2881 Make all REST request un-cached.
IE was caching some REST requests (we didn’t say they weren’t) so using old data. We now say that all the JSON data returned in uncacheable so it has to be re-retrieved each time. The means that when a signup is added by an administrator the top table of the administration page is also updated with the new data.