WL-4029 Allow exceptions of text/xml resources to work.
When we were mapping the exceptions onto responses we didn’t specify a type and it was attempting to map them to the requested response type (text/xml), as we have no MessageBodyWriters for this type it would fail. As we only have MessageBodyWriters for application/json we force all mapped exceptions to be of this type.