WL-548 Moving some message properties to an i8n aware location.
For some reason, certain properties related to the page which adds
participants to a site, expected to find these values in sakai.properties.
I have moved them to the (more correct) sitesetupgeneric.properties which
makes them amenable for localization in to other languages.

WL-869 Don't escape some HTML in a bundle.
WL-869 Reinstated the lost link to university contact search (add participants).

The message bundles have become an unholy mess :-( - central Sakai, not us! In particular, many of them are found in similar (but differing) forms under one of
the following directories:

site-manage-tool/
site-manage-participant-helper/

It appears that the latter is the new correct location. My original commit for
this involved changing the names of the properties too. I have removed all
references to these and used the original ones. As the values are appropriate
for either locale (en_US & en_GB) I have placed them in the default message
bundle.

SAK-17225 Switch to RSF way.

Conflicts:
	site-manage-participant-helper/src/bundle/org/sakaiproject/site/tool/participant/bundle/sitesetupgeneric.properties