WL-720 Added a name-value pair for GeneralActionError to poll messages.
GeneralActionError & GeneralShowError are name-value pairs expected to be
found by the RSF framework. I have added both for the sake of completeness
and based their values on their corresponding values in evaluation (another
RSF tool). As an interim solution I have appended a warning to do
with the additional instructions for the poll being too long. A better
more complete solution is to change the type for the POLL_DETAILS field in
the POLL_POLL table (see SAK-13310), in order to overcome the current 255
character limit.