WL-900 Preserve GET parameters through login
We currently drop any query parameters when redirecting to login. The
same fix was recently applied to the access tool, now we are applying it
here too.

This is particularly important if we want to remove breadcrumbs through

    ?breadcrumb=0

The method for constructing the return URL is directly copied from access:

https://github.com/ox-it/wl-access/pull/3