WL-3676 Only let people who can use tool see it.
Anonymous users can’t use the signup tool as it uses the SessionManager to get IDs which results in nulls for the anonymous user. Switching to the UserDriectoryService which returns empty strings is a little risky and would need lots of testing to so mask the issue we’ll just hide the tool from those users won’t can’t use it.