WL-3561 Allow topic creation by /direct.
In the upgrade to Sakai 10 the getting of the current user now needs to know the current site, so we have to add the siteId through to all these calls.

WL-3561 Saving of a message when values aren’t passed

We now are much more robust in how we work out values when they aren’t passed to us. This patch is far from ideal as it means there are multiple ways to work the information out, but it allows the code to work for both the tool (without change) and from /direct.

WL-3561 Don’t assume all exceptions are security.

Catching all exceptions and throwing them as a security problem is just wrong and hides the underlying cause as entity broker doesn’t logs anything much when there is a security problem as it assumes that it’s to be expected.