WL-7 Hierarchy Changes - /tetraelf-hierarchy
This switches to use colons to encode the parts of the hierarchy. It also switches to use the new SiteNeighbourhoodService in the portal so that we don’t have specific handlers for all the portal calls.

The old URLs have redirects to the new locations so existing links should work.

Also update the webapp for Sakai 11.