WL-3027 Drop old search from the build.
Having both copies of search in the build resulted in too many EntityContentProducers getting registered and it not being predictable which one would get used. As we have duplicates of the 2 in search-impl in solr-impl we can remove it all from our build. It wasn't consistent which ECP got found so indexing would produce unpredictable results. Ideally all the ECPs should be in search-adaptors but that's for when we do a larger refactor.