Solr implementation.
This is a squash of a large number of commits to make merging easier.

WL-2871
WL-3027 Don't return contents in search results.
WL-3032 Create 2 solrservers with different timeouts.
WL-3074 When doing a full re-index nothing was looked at.

The content producer wasn't getting called when doing a full re-index as it was checking the reference against the list of accepted types. Example of the problem of using Strings everywhere, enums help prevent problems like this.

Sakai 10 drops id and subtype.

Sakai 10 no longer has ID and subtype fields, so we no longer attempts
to put these values into the index. We also update the schema to remove
these fields.

WL-3356 Don’t leak memory when indexing PDFs.

PDFBox keeps a cache of something todo with fonts and over time this just accumulates heap. To work around this we clear out the cache after indexing each document. I’ve also bumped up the version of PDFBox to the latest version.

WL-3489 Update tika from 1.3 to 1.5

WL-3424 Don’t warn about not indexing directories.

WL-749 Extract the filename and include as prop.

The solr config will look for a field of property_filename and if it exists copy it into the filename field. We don’t put it directly into the filename field as the properties are the only way to pass additional data to the index.

The solr config will also include the filename field in all searches.

WL-3612 Don’t every return null as it NPEs.

The solr implementation of search was returning a null for suggestions which was causing the search tool to NPE. Now we just return an empty array.