WL-3068 Force IE to parse xml using its own special way
Currently IE gets tripped up when getting the RSS XML from MediaPub, this
results in the success callback never being fired. As a result the podcast
picker would never open.

The fix is to get jQuery to load the feed as if it is plain text and then
get IE to parse the XML as part of the success callback.