Give tools something to render
There are approximately 890 references to /library/skin/default/tool.css in
the various html files which are dotted around the place, despite the default
folder not existing any more.

As a quick fix, I am copying the tool.css file from our current stylesheet
so that it has something to serve up.