WL-2956 Do the re-writing of content when parsing.
This removes unused JavaScript. If you attempt to fix YouTube URLs in IE9 by re-writing the DOM using a DOMContentLoaded or onload event it doesn’t fire and you just get a warning about insecure content. Switching back to rendering the page as a IE8 Document works, but we don’t want to force all modern versions of IE to be crippled. The only other fix that worked was to actually re-write the content when the page is loading. This still results in a warning message, but it does at least allow the page to work.