WL-3714 : Fix IE11 'skip' text so it renders offscreen
The 'Minimise' em text (used for accessibility) which has position:fixed is as a result
given display:block in Firefox and Chrome and IE 8,9 but not in IE11
which has display:inline.

Changing it to display:block does not make a difference.
This change moves the text offscreen.

WL-3714 : Move -999px from tool_base.css to portal.css