WL-4542 Switch to VARCHAR rather than INTs for enum.
This makes the oauth code use VARCHARs rather than INTs for it’s enums as it makes the database more readable and helps to prevent problems when enums are changed.

This is also needed as we currently have this setup from our 10.x OAuth deployment and migrating to INTs seems like a backwards step.