WL-4539 Don’t fail elfinder when can’t access site.
We have secure sites which you can only access once you are 2 factor authenticated and so failing the whole connector request if one of the sites the user is a member of isn’t accessible is wrong. This is a 2 part fix, firstly we filter out the sites that can’t currently be accessed, secondly if someone does attempt to access one we just show it as empty.