WL-4634 Corrected formatting for Lessons forums widget.
Adjusted css for forum widget so that it always fits in the container.Have
merged latest changes from sakai too like 'change in the color of widget
header' and html for forum-summary type item will be generated on the fly
instead of storing it in database.
Also have added 'forumSummary-addBefore' param for the widget so that it
can be added at any place inside Lessons instead of getting added at
the end of the page everytime.