WL-4633 Corrected formatting for Lessons announcements widget.
Adjusted css for announcements widget so that it always fits in the
container.Have merged latest changes from sakai too like 'change in
the color of widget header' and html for announcements type item
will be generated on the fly instead of storing it in database.
Also have added 'announcements-addBefore' param for the widget so
that it can be added at the correct place on '+' click instead
of getting added at the bottom of the page all the times.