WL-4656 Allow maintainers to add LTI tools.
In Sakai 10 maintainers in a site could add an LTI tool to their site one or more times, in Sakai 11 this is only available to the admin user in the administration workspace. This adds it back and controls it with a property, which defaults to true, to stop maintainers adding multiple tools add:

    lti:allow_maintainer_add_tool_site=false

to sakai.properties.