WL-3255 Don’t allow deletion of !hierarchy.missing
This also means that we display a better message when the user can’t delete the site but can remove the node. However the terminology is still confusing and could do with further work. This is a belt and braces solution which updates the template to stop users getting the option but also updates the server side code to do the check there as well.