WL-4505 Keep property updated on managed sites.
When a admin site is changed on a site the property on the site is updated. Also when a title of an admin site is changed all the sites managed by that site are updated if they don’t have the property set correctly.

Also fixed a bug where unmanaging a site would result in it being managed by a non-existent site. We also fire an event when a site becomes unmanaged.