WL-4504 Fix for NPE when updating admin sites.
Because all the sites didn’t have an existing property set when attempting to generate an admin site update event the event listener failed with a NPE. This changes means that it works correctly even if you don’t already have a admin site property set.

I’ve also updated the backfill job so you can specify the event as this allows you to generate the admin change event across all project sites which would have also allowed the update to work, but the currently deployed version doesn’t have this so we couldn’t.