SAK-32131 Fix issue with lowercase growing string. (#3827)
* SAK-32131 Fix issue with lowercase growing string.

Some characters get longer when they become lowercased and this was causing the portal code to incorrectly inline the tool. In severe cases this would cause an out of bounds exception.

In other cases the returned page would contain multiple closing body tags.
The refactor here makes it possible to write a unit test for the code without having to mock large parts of the request.

* Clean up debug statement to use substitution

 Conflicts:
	portal/portal-impl/impl/pom.xml