WL-4853 Allow import from site on custom site type.
When you are in a site that is inheriting it’s tools from another site type through configuration:

    # This allows repository site types to have all the normal project tools.
    projectSiteType=repository,project
    projectSiteTargetType=project

The tools list presented is correct, but when you attempt to perform the import it fails. This is because when checking that the tools are valid for the site being imported into it doesn’t use this configuration to locate the list of valid tools.

Although this isn’t an API I’ve added some tests to that we can check that the internal methods are behaving as expected.