WL-4786 Fix replace site to work again.
This was broken because there are special tool paths that now get broken and this includes anything starting /site (ahhh). So we change the internal tool URL we are using. Also we were redirecting to the wrong URL, now we redirect back to the current location in the hierarchy, showing the new site in place.