SAK-32398 Remove uses of byte array for uploads (#4181)
* SAK-32398 Remove/correct JavaDoc.

These files are touched by this change and have JavaDoc that is wrong.

* SAK-32398 Remove code for reading uploads to heap

This is so that we don’t have to worry about heap when/if we
enable large file uploads and someone finds a request that stupidly
puts it all on the heap.

Where possible we switched to streams. AttachmentAction isn't used at
all so I removed this class and moved the only used constant out to the
class that uses it.

The Sakai FileItem no longer allows you to get the byte array of the
upload and the ParameterParser doesn't fallback to using a byte array,
but the input stream should work all the time.

The removed section in FilePickerAction.java is no longer needed as it
was there for WebSphere and uploads in old gradebook work without it.

* SAK-32398 Log when we fail to get an inputstream.