SAK-32360 Always pass site ID to LTI calls. (#4134)
* SAK-32360 Always pass site ID to LTI calls.

The LTIService now always wants the site ID passed into all it’s calls as assuming that the context is correct doesn’t work when the tool is in helper mode in My Workspace or the Administration Workspace.

As lots of the method signatures were changing I also did a cleanup, removing lots of useless JavaDoc and adding @Override to the appropriate methods.

* SA-32360 Update calls in .vm files to LTIService.

The velocity template files were making calls directly into the service and this didn’t get caught be the refactoring.

* SAK-32360 Removed unused/unimplemented methods.

Some methods were still in the API which because of the refactor were being mistakenly called instead of the correct ones, these have now been removed.

* SAK-32360 Old code didn’t appear to actually work.

* SAK-32360 Correct typo

* SAK-32360 Correct Lessonbuilder BltiEntity.

In lesson builder the current API doesn’t make it easy to work out the site ID so just using the current placement.

* SAK-32360 When creating a site we don’t have a ID

So instead use a random ID.

* SAK-32360 Fix for missing import in SiteAction

WL-4744