WL-4926 Switch from WebAuth to Shibboleth.
We are using the Apache module for shibboleth logins. This means changing out apache docker image. We now install the shibboleth code. We then have to have an additional docker container that runs the shibd daemon. The shied daemon and apache communicate through a socket on a shared volume, this means that they both need to be running on the same host, the volume also has to be a unix volume. There is the option of using a TCP connection and this looks to work, but then requires config re-writing to embed the IP of the endpoint.

This update also switches to docker compose version 2 as we needed cleaner shared volumes and this adds the ability to have the apache config generated in a better way by having a DNS alias that points to all the workers.

You need to set the server name in docker/sakai/.env so that apache knows it’s hostname. You also need to generate a shibboleth key pair for use with talking to the federation.

Set the logout URL when doing shibboleth, at the moment we just do local
logout.

We have the shibboleth module de-scope the username so that
buckett@OX.AC.UK becomes buckett which means we don't have todo this in
the Java code. Otherwise we would have to provide a way todo searches
based on the scoped username and a way todo searches on the un-scoped
username so that people could still search based on just the username.