SAK-32561 Move the smileys to a fixed location. (#4447)
* SAK-32561 Move the smileys to a fixed location.

This way when ckeditor gets upgraded we don’t lose the icons and people don’t end up with broken images in the HTML. This does have the downside that new images won’t get picked up automatically and any changes to icons won’t get picked up automatically.

* SAK-32561 Forward creditor smiley images.

We send all requests to the current webjar path.

We have an additional rule that at the moment is redundant but when ckeditor is next upgraded will mean that the old links continue to work. Once ckeditor has been upgraded we can add FORWARD back to the filter dispatchers that are used, but at the moment while the versions are the same it results in a endless loop.