WL-4964 Fix for course signup rendering badly.
With course signup moving out of an iframe on Sakai 11 it causes the whole of Sakai to render badly in IE11. The reason is that it forces Sakai to render everything in IE 8 mode which doesn’t work at all. Switching to render in IE11 mode looks to generally be better although there maybe some quirks in edge cases.

However this change shouldn’t affect anything other than IE so it should be reasonably low risk for everyone else.