WL-4976 Remove custom hibernate index creation.
We rely instead on hibernate creating indexes itself when DDL is set. The name of the pathHash index changes, but this doesn’t appear to be an issue. This gets rid of some errors in the logs that weren’t actually errors but were expected when starting up against an existing database.