SAK-31980 poi 3.15, Signup poi 3.14 (#4182)
The pattern for creating a Hyperlink has changed so that now you use the creation helper associated with the workbook to create the links. This means signup can stay in sync with POI in the rest of Sakai.

(cherry picked from commit 57ecb6bfcc1628a13dc34528ab0f49f22b7b2e16)