SAK-32389 Restoring behavior of tool menu toggle button (#4518)
Before commit 3103b77 clicking anywhere in the <li> for the collapse button would toggle the menu. Changes made in that commit made it so that the double arrows needed to be clicked directly. This commit is an attempt to restore the previous behavior and preserve the changes made in 3103b77.