Fix for failure to start Apache.
Sometimes Apache would get killed and wouldn’t remove it’s PID file. This would prevent apache from starting up again and the only way to get it to start again would be to remove the container. This checks to see if the PID file is there at startup and if it is removes it.