WL-4917 Add support for passing anonymous to TII.
TII has an anon parameter that accepts 1 or 0 and is used to indicate if it should hide the user’s names (making it anonymous). As we have support for anonymous assignments inside Sakai we can pass across this value to TII.

When testing this the TII account needs to have the anonymous feature enabled. An error is not generated if you attempt to setup an anonymous survey when it’s not enabled.

The anonymous flag isn’t on the AssignmentContent so we have to pass it down through the call stack like the open/close dates. I wonder if it should have been in the AssignmentContent rather than the Assignment Properties, but refactoring this isn’t trivial.