WL-5046 Make sure logging for current-load is seen
The current load servlet should log as an error when we hit the sleep
limit as we should be reconfiguring things if this is getting hit as it
means that our sessions won't be well balanced.

Also we want info logging for it.ox.ac.uk which should have got enabled
earlier. Also code tidy ups in the load servlet.