WL-5060: Students cannot resubmit in an AS site
When the list of assignments is displayed, the link on the assignment title goes to one vm if resubmits are allowed and another vm if they are not. The cause of this bug is that it is incorrectly calculating that AS assignments are not resubmittable and linking to the wrong vm (a vm where you cannot resubmit despite the fact that students still want to).

The test for whether an assignment is resubmittable or not is (ridiculously) whether there are previous submissions and whether the content of the log for any of those previous submissions contains the word 'submitted'. Unfortunately, when an AS assignment is submitted to, the word 'submitted' is not added in - which I have changed. Flaaaaaky.