SAK-33561 Support for filter custom properties. (#4960)
* SAK-33561 Support for filter custom properties.

This allows local deployments to customise the LTI substitution
properties that are used for custom property subsitution.

* SAK-33561 Added example substitution filter.

This substitution filter adds a random number to the substituted variables. You can then use this in your custom parameters with:

    lucky=$random

This filter is only enabled when Sakai is started in demo mode (with the JVM flag of -Dsakai.demo=true).

* Added some more JavaDoc.