WL-5088 Fix for Daisy courses not being deleted.
When the Daisy/TMS import was being run the courses that no longer appeared in the feed were no longer being deleted. The reason for this was that although a transaction proxy was setup on the wrapper the method being called didn’t have the transaction proxy managing it. The call to the method that did have the transaction proxy was done internally to the class and so Spring’s AOP proxy didn’t intercept the call and so the transaction wasn’t applied.

I think this also affected the OxCap importers that should delete everything but this wasn’t noticed because these feeds don’t change as much.

I’ve refactored things a little in this fix so that we get rid of the PopulatorWrapper as it was redundant and just have everything implement the Populator interface.