WL-5094 When anonymous don’t release marks for year
If an assignment in Sakai is anonymous and the submissions are being sent to Turnitin. Tell Turnitin not to release the marks/feedback for a year. This is done because when the marks are released the anonymity is removed from the Turnitin reports.

We don’t do this in the case of a non-anonymous assignment incase the instructor is providing feedback in Turnitin and we don’t want to delay feeding back for a year. The alternative would have been to expose this information in Assignments and allow the assignment creator to choose when to release the marks/anonymity.