WL-5182 Use placement advice in Content Item Message
When a Content Item Message is received it can advise about how the item should be displayed. This wasn’t being followed and Sakai was just producing a link to the selected item.

Now we follow it as best we can and fallback to just showing a link if there isn’t any advice.

We also now generate a relative URL so that the antisamy policy doesn’t have to include the current host. Using a relative hostname also means that links work when databases are duplicated to test/staging systems, rather than having them point back to the original system.