WL-5195 Don’t log GradebookNotFoundExceptions
When grademark syncing is enabled it tries to sync the grades even if the
gradebook isn’t on the site. This checks first that there is a gradebook.
It updates the security advisors to use the improved push/pop.
It improves the try/catch so that we don’t log and error and then NPE on
a different line, but catch the first error and don’t attempt to continue.

trademark is enabled by default.