WL-5202 Correctly pull through forename/surname.
When the forename and surname were set on the user object but weren’t set on the ValidationAccount object the value would be shown in the form but then when submitted it would generate an error saying that the forename was empty.

This was because the value wasn’t set on the bean backing the form and RSF only updates values that are considered too have changed. The 2 ways this can be solved are:

  - force the UIInputs to always update the bean. This means setting `mustapply` in the producer and then the default values are re-applied to the bean when the form data is read in even though the new value and the fossil values are the same. This isn’t optimal because we’re fixing it late in the call.
  - update the ValidationAccount bean with defaults from the user object if they aren’t set on the ValidationAccount itself. This is the cleaner solution as the ValidationAccount bean correctly represents the state of the form.