WL-5221 Refactoring of external groups.
This updates the external groups code to use the connection pool directly so that it doesn’t result in closing connections.

This also pulls all the custom logic out of the handlers and puts it in the external group manager, this means we have less if (unit) … if (course) … chunks of code as we can pass in the correct chunk to the actual function.

We also bump to up project to Java 1.8.