WL-5226 Fix for the sakai user not working.
When running tomcat as the Sakai user (not root) it would fail to startup because the user didn’t have permissions to read the tomcat files. I’ve fixed this by making all the tomcat files owned by the sakai group.

I also moved the section later in the dockerfile so that more of the network related stuff happens earlier and makes testing easier.