WL-5249 Don’t remove all the provided members.
When updating a group’s members don’t remove all the provided members. Just leave them and then when the membership is refreshed by the AuthzGroupService all the provided members will get updated.

This does mean that if you remove a section from a group it won’t immediately update, but also means that if edit a group with just sections in it the membership doesn’t jump to zero on saving.