WL-5226 Fix the RTT to work with Tomcat 8.5.x
We were attempting to use a request dispatcher to forward the request to //rest//search/.... and this was returning a 404. The problem was the double slash at the start of the path. This appeared to “work” under earlier versions of Tomcat.

This was happening because the Spring view resolver was prepending "/" to all views. The fix was to stop doing this and to stop trimming off the leading slash in the controller to generate the view.