WL-5260 Fix for character encoding in JVM
The JVM wasn’t being correctly set to UTF-8 and this was causing the
portal inlining to use ASCII, then then meant that when a unicode non
breaking space was written out it mapped to a ? rather than a non
breaking space (160).

It’s Antisamy that was converting the &nbsp; to unicode 160, but we
haven’t changed any of that translation in the commit.

Rather than reconfiguring the locales package after installing we answer
the questions before installing the package.